+++ posthog-samples/posthog-android-sample/src/main/java/com/posthog/android/sample/DoSomething.kt
@@ -1,0 +1,13 @@
+package com.posthog.android.sample
+
+import com.posthog.PostHog
+
+class DoSomething {
+    fun doSomethingNow() {
+        try {
+            throw MyCustomException("Something went wrong")
+        } catch (e: Throwable) {
+            PostHog.captureException(e, mapOf("am-i-stupid" to true))
+        }
+    }
+}