+++ posthog/src/main/java/com/posthog/PostHog.kt
@@ -19,6 +19,7 @@ import com.posthog.internal.PostHogSendCachedEventsIntegration
 import com.posthog.internal.PostHogSerializer
 import com.posthog.internal.PostHogSessionManager
 import com.posthog.internal.PostHogThreadFactory
+import com.posthog.internal.exceptions.ThrowableCoercer
 import com.posthog.internal.replay.PostHogSessionReplayHandler
 import com.posthog.internal.surveys.PostHogSurveysHandler
 import com.posthog.vendor.uuid.TimeBasedEpochGenerator
@@ -61,6 +62,8 @@ public class PostHog private constructor(
 
     private var isIdentifiedLoaded: Boolean = false
     private var isPersonProcessingLoaded: Boolean = false
+
+    private val throwableCoercer = ThrowableCoercer()
 
     // this is called if the feature flags are loaded for the first time and recording isn't started yet
     private val internalOnFeatureFlagsLoaded =
@@ -477,6 +480,38 @@ public class PostHog private constructor(
             }
         } catch (e: Throwable) {
             config?.logger?.log("Capture failed: $e.")
+        }
+    }
+
+    override fun captureException(
+        throwable: Throwable,
+        properties: Map<String, Any>?,
+    ) {
+        if (!isEnabled()) {
+            return
+        }
+
+        try {
+            val exceptionProperties =
+                throwableCoercer.fromThrowableToPostHogProperties(
+                    throwable,
+                    inAppIncludes = config?.inAppIncludes ?: listOf(),
+                )
+
+            properties?.let {
+                exceptionProperties.putAll(it)
+            }
+
+            config?.let { config ->
+                val personUrl = "${config.host.trimEnd('/')}/project/${config.apiKey}/person/${distinctId()}"
+                exceptionProperties["\$exception_personURL"] = personUrl
+            }
+
+            capture(PostHogEventName.EXCEPTION.event, properties = exceptionProperties)
+        } catch (e: Throwable) {
+            // we swallow all exceptions that the SDK has thrown by trying to convert
+            // a captured exception to a PostHog exception event
+            config?.logger?.log("captureException has thrown an exception: $e.")
         }
     }
 
@@ -1088,6 +1123,13 @@ public class PostHog private constructor(
                 userPropertiesSetOnce = userPropertiesSetOnce,
                 groups = groups,
             )
+        }
+
+        public override fun captureException(
+            throwable: Throwable,
+            properties: Map<String, Any>?,
+        ) {
+            shared.captureException(throwable, properties)
         }
 
         public override fun identify(