+++ posthog-android/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Next
 
-## 3.24.0 - 2025-10-09
+- feat: add manual captureException ([#300](https://github.com/PostHog/posthog-android/issues/300))
+
+## 3.24.0 - 2025-10-10
 
 - add evaluation tags to android SDK ([#301](https://github.com/PostHog/posthog-android/pull/301))